+++ mirrord-cli/src/main.rs
@@ -108,8 +108,7 @@ fn exec(args: &ExecArgs) -> Result<()> {
 
     if let Some(pod) = &args.pod_name {
         // TODO: do we need a print here or just a log is fine?
-        println!("[WARNING]: DEPRECATED - `--pod-name` is deprecated, consider using `--target instead.\nDeprecated since: [24/09/2022] | Scheduled removal: [24/10/2022]");
-        warn!("[WARNING]: DEPRECATED - `--pod-name` is deprecated, consider using `--target instead.\nDeprecated since: [24/09/2022] | Scheduled removal: [24/10/2022]");
+        println!("[WARNING]: DEPRECATED - `--pod-name` is deprecated, consider using `--target instead.\nDeprecated since: [28/09/2022] | Scheduled removal: [28/10/2022]");
         std::env::set_var("MIRRORD_AGENT_IMPERSONATED_POD_NAME", pod);
     }
 
@@ -192,6 +191,10 @@ fn exec(args: &ExecArgs) -> Result<()> {
 
     if args.no_outgoing || args.no_udp_outgoing {
         std::env::set_var("MIRRORD_UDP_OUTGOING", "false");
+    }
+
+    if let Some(config_file) = &args.config_file {
+        std::env::set_var("MIRRORD_CONFIG_FILE", config_file.clone());
     }
 
     let library_path = extract_library(args.extract_path.clone())?;